backend/requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.0
Flask-JWT-Extended==4.4.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.0
bcrypt==4.0.0
pydantic==1.10.0
PyPDF2==3.0.0
reportlab==4.0.0
Werkzeug==2.3.0
marshmallow==3.19.0
pytest==7.0.0
pytest-cov==4.0.0
